armsoc_dri2: Also swap buffers on partial blits
When a direct swap cannot be accomplished, perhaps because the window is
partially obscured or hidden, Mali still expects the buffers to have
been swapped from its point of view and won't requery with
GetBuffers. Make it happy by always swapping the front and back, even if
we did a blit.